feat: Add database migration Init
Add the Init migration, effectively creating the database.

In the CourseStudent table, I renamed the auto-generated CoursesId and StudentsId to CourseId and StudentId. I think this is more in-line with their references to the original tables, which also aren't pluralized. However, it may be a convention for many-to-many tables, in which case the renaming will have to be undone.

The web project now compiles and displays pages.