feat: Add Enrollment entity
Add the Enrollment entity, which has both a Course and a Student as its primary key. Add the relationship to the Enrollment into the Course and Student classes.

To add this to the database, the LexiconUniversityContext requires an OnModelCreating method. In this method the many-to-many relationship of the Enrollment table is defined.