feat: Add Enrollment table to DbContext
Add Enrollment as an entity to DbContext. Include the LexiconUniversityContextFactory and its configuration to be able to create the relevant migration and database update.

With these changes, VS could scaffold an EnrollmentsController. This was not the original goal but a fine test of design-time changes to an "external" (to the application project) database.