feat: Create StudentSummaryViewModel for Students.Index view
Create the StudentSummaryViewModel which includes only the Id, Avatar, FullName, and Email of the student for displaying on the Index (list) page.

I tried to make this into an efficient SQL query by selecting only the necessary columns to create the type. My next step is to try to add pagination to this.