feat: Add AutoMapper
Add the AutoMapper, a tool for automating the transition from one class to another, in this case between the Entities and the Views. Apply the mapper to creating the StudentSummaryViewModel for Students/Index.

In its current form this creates a complete Select statement in the resulting SQL. There are "Queryable Extensions" for AutoMapper that should only select the required columns but I couldn't get this to work yet.